Removed now unused stuff.  Will be implementing replacements soon.
Will also need to re-fill documentation slightly. Although, I hope it
needs less documentation as things should be clearer, more straightforward,
and more akin to things people already know (of net/http for ex.)
