+++ https.go
@@ -10,35 +10,8 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"strconv"
 	"strings"
-	"sync/atomic"
-
-	"github.com/inconshreveable/go-vhost"
 )
-
-type ConnectActionLiteral int
-
-const (
-	ConnectAccept = iota
-	ConnectReject
-	ConnectMitm
-	ConnectHijack
-	ConnectHTTPMitm
-)
-
-var (
-	OkConnect       = &ConnectAction{Action: ConnectAccept, TLSConfig: TLSConfigFromCA(GoproxyCa)}
-	MitmConnect     = &ConnectAction{Action: ConnectMitm, TLSConfig: TLSConfigFromCA(GoproxyCa)}
-	HTTPMitmConnect = &ConnectAction{Action: ConnectHTTPMitm, TLSConfig: TLSConfigFromCA(GoproxyCa)}
-	RejectConnect   = &ConnectAction{Action: ConnectReject, TLSConfig: TLSConfigFromCA(GoproxyCa)}
-)
-
-type ConnectAction struct {
-	Action    ConnectActionLiteral
-	Hijack    func(req *http.Request, client net.Conn, ctx *ProxyCtx)
-	TLSConfig func(host string, ctx *ProxyCtx) (*tls.Config, error)
-}
 
 func stripPort(s string) string {
 	ix := strings.IndexRune(s, ':')
@@ -60,223 +33,6 @@ func (proxy *ProxyHttpServer) connectDial(network, addr string) (c net.Conn, err
 		return proxy.dial(network, addr)
 	}
 	return proxy.ConnectDial(network, addr)
-}
-
-func (proxy *ProxyHttpServer) handleHttpsConnect(w http.ResponseWriter, r *http.Request) {
-	ctx := &ProxyCtx{Req: r, Session: atomic.AddInt64(&proxy.sess, 1), proxy: proxy}
-
-	hij, ok := w.(http.Hijacker)
-	if !ok {
-		panic("httpserver does not support hijacking")
-	}
-
-	preTLSClient, _, e := hij.Hijack()
-	if e != nil {
-		panic("Cannot hijack connection " + e.Error())
-	}
-
-	preTLSClient.Write([]byte("HTTP/1.0 200 OK\r\n\r\n"))
-	proxyClient, vhostErr := vhost.TLS(preTLSClient) // try to sniff SNI
-	if vhostErr != nil {
-		ctx.Logf("Failed to sniff SNI (falling back to request Host): %s", vhostErr)
-	}
-
-	ctx.Logf("Running %d CONNECT handlers", len(proxy.httpsHandlers))
-	host := proxyClient.Host()
-	if host == "" {
-		host = r.URL.Host
-	}
-	todo := OkConnect
-	for i, h := range proxy.httpsHandlers {
-		newtodo, newhost := h.HandleConnect(host, ctx)
-
-		// If found a result, break the loop immediately
-		if newtodo != nil {
-			todo, host = newtodo, newhost
-			ctx.Logf("on %dth handler: %v %s", i, todo, host)
-			break
-		}
-	}
-	switch todo.Action {
-	case ConnectAccept:
-		if !hasPort.MatchString(host) {
-			host += ":80"
-		}
-		targetSiteCon, err := proxy.connectDial("tcp", host)
-		if err != nil {
-			proxyClient.Close()
-			//httpError(proxyClient, ctx, err)
-			// .. previously, we would return
-			// an appropriate error if we weren't able to connect to the remote
-			// node.
-			return
-		}
-		ctx.Logf("Accepting CONNECT to %s", host)
-		go copyAndClose(ctx, targetSiteCon, proxyClient)
-		go copyAndClose(ctx, proxyClient, targetSiteCon)
-	case ConnectHijack:
-		ctx.Logf("Hijacking CONNECT to %s", host)
-		todo.Hijack(r, proxyClient, ctx)
-	case ConnectHTTPMitm:
-		ctx.Logf("Assuming CONNECT is plain HTTP tunneling, mitm proxying it")
-		targetSiteCon, err := proxy.connectDial("tcp", host)
-		if err != nil {
-			ctx.Warnf("Error dialing to %s: %s", host, err.Error())
-			return
-		}
-		for {
-			client := bufio.NewReader(proxyClient)
-			remote := bufio.NewReader(targetSiteCon)
-			req, err := http.ReadRequest(client)
-			if err != nil && err != io.EOF {
-				ctx.Warnf("cannot read request of MITM HTTP client: %+#v", err)
-			}
-			if err != nil {
-				return
-			}
-			req, resp := proxy.filterRequest(req, ctx)
-			if resp == nil {
-				if err := req.Write(targetSiteCon); err != nil {
-					httpError(proxyClient, ctx, err)
-					return
-				}
-				resp, err = http.ReadResponse(remote, req)
-				if err != nil {
-					httpError(proxyClient, ctx, err)
-					return
-				}
-			}
-			resp = proxy.filterResponse(resp, ctx)
-			if err := resp.Write(proxyClient); err != nil {
-				httpError(proxyClient, ctx, err)
-				return
-			}
-		}
-	case ConnectMitm:
-		ctx.Logf("Assuming CONNECT is TLS, mitm proxying it")
-		// this goes in a separate goroutine, so that the net/http server won't think we're
-		// still handling the request even after hijacking the connection. Those HTTP CONNECT
-		// request can take forever, and the server will be stuck when "closed".
-		// TODO: Allow Server.Close() mechanism to shut down this connection as nicely as possible
-		tlsConfig := defaultTLSConfig
-		if todo.TLSConfig != nil {
-			var err error
-			tlsConfig, err = todo.TLSConfig(host, ctx)
-			if err != nil {
-				httpError(proxyClient, ctx, err)
-				ctx.Logf("Couldn't configure TLS: %s", err)
-				return
-			}
-		}
-		go func() {
-			//TODO: cache connections to the remote website
-			rawClientTls := tls.Server(proxyClient, tlsConfig)
-			if err := rawClientTls.Handshake(); err != nil {
-				ctx.Warnf("Cannot handshake client %v %v", r.Host, err)
-				return
-			}
-			defer rawClientTls.Close()
-
-			ctx.Conn = rawClientTls
-			clientTlsReader := bufio.NewReader(rawClientTls)
-			for !isEof(clientTlsReader) {
-				req, err := http.ReadRequest(clientTlsReader)
-				if err != nil && err != io.EOF {
-					return
-				}
-				if err != nil {
-					ctx.Warnf("Cannot read TLS request from mitm'd client %v %v", r.Host, err)
-					return
-				}
-				req.RemoteAddr = r.RemoteAddr // since we're converting the request, need to carry over the original connecting IP as well
-				ctx.Logf("req %v", r.Host)
-				req.URL, err = url.Parse("https://" + r.Host + req.URL.String())
-
-				ctx.proxy.dispatchRequestHandlers(ctx)
-				// req, resp := proxy.filterRequest(req, ctx)
-				// if resp == nil {
-				// 	if err != nil {
-				// 		ctx.Warnf("Illegal URL %s", "https://"+r.Host+req.URL.Path)
-				// 		return
-				// 	}
-				// 	removeProxyHeaders(ctx, req)
-				// 	resp, err = ctx.RoundTrip(req)
-				// 	if err != nil {
-				// 		ctx.Warnf("Cannot read TLS response from mitm'd server %v", err)
-				// 		return
-				// 	}
-				// 	ctx.Logf("resp %v", resp.Status)
-				// }
-				// resp = proxy.filterResponse(resp, ctx)
-				resp := ctx.Resp
-
-				text := resp.Status
-				statusCode := strconv.Itoa(resp.StatusCode) + " "
-				if strings.HasPrefix(text, statusCode) {
-					text = text[len(statusCode):]
-				}
-				// always use 1.1 to support chunked encoding
-				if _, err := io.WriteString(rawClientTls, "HTTP/1.1"+" "+statusCode+text+"\r\n"); err != nil {
-					ctx.Warnf("Cannot write TLS response HTTP status from mitm'd client: %v", err)
-					return
-				}
-				// Since we don't know the length of resp, return chunked encoded response
-				// TODO: use a more reasonable scheme
-				resp.Header.Del("Content-Length")
-				resp.Header.Set("Transfer-Encoding", "chunked")
-				if err := resp.Header.Write(rawClientTls); err != nil {
-					ctx.Warnf("Cannot write TLS response header from mitm'd client: %v", err)
-					return
-				}
-				if _, err = io.WriteString(rawClientTls, "\r\n"); err != nil {
-					ctx.Warnf("Cannot write TLS response header end from mitm'd client: %v", err)
-					return
-				}
-				chunked := newChunkedWriter(rawClientTls)
-				if _, err := io.Copy(chunked, resp.Body); err != nil {
-					ctx.Warnf("Cannot write TLS response body from mitm'd client: %v", err)
-					return
-				}
-				if err := chunked.Close(); err != nil {
-					ctx.Warnf("Cannot write TLS chunked EOF from mitm'd client: %v", err)
-					return
-				}
-				if _, err = io.WriteString(rawClientTls, "\r\n"); err != nil {
-					ctx.Warnf("Cannot write TLS response chunked trailer from mitm'd client: %v", err)
-					return
-				}
-			}
-			ctx.Logf("Exiting on EOF")
-		}()
-	case ConnectReject:
-		if ctx.Resp != nil {
-			if err := ctx.Resp.Write(proxyClient); err != nil {
-				ctx.Warnf("Cannot write response that reject http CONNECT: %v", err)
-			}
-		}
-		proxyClient.Close()
-	}
-}
-
-func httpError(w io.WriteCloser, ctx *ProxyCtx, parentErr error) {
-	ctx.Logf("Sending http error: %s", parentErr)
-	if _, err := io.WriteString(w, "HTTP/1.1 502 Bad Gateway\r\n\r\n"); err != nil {
-		ctx.Warnf("Error responding to client: %s", err)
-	}
-	if err := w.Close(); err != nil {
-		ctx.Warnf("Error closing client connection: %s", err)
-	}
-}
-
-func copyAndClose(ctx *ProxyCtx, w, r net.Conn) {
-	connOk := true
-	if _, err := io.Copy(w, r); err != nil {
-		connOk = false
-		ctx.Warnf("Error copying to client: %s", err)
-	}
-	if err := r.Close(); err != nil && connOk {
-		ctx.Warnf("Error closing: %s", err)
-	}
 }
 
 func dialerFromEnv(proxy *ProxyHttpServer) func(network, addr string) (net.Conn, error) {
